Fixed an Error with crashing DE16x18 Layout

+++ include/led.hpp
@@ -318,7 +318,7 @@ inline void Led::clearRow(uint8_t row) {
 
 inline void Led::clearMinArray() {
     if (usedUhrType->numPixelsFrameMatrix() == 0) {
-        for (uint8_t i = usedUhrType->numPixelsWordMatrix();
+        for (uint16_t i = usedUhrType->numPixelsWordMatrix();
              i < usedUhrType->numPixels(); i++) {
             clearPixel(i);
         }
